Refactor workflow management and enhance plan linking capabilities
- Introduced the PlanCommitService to handle automatic commits for plan phases, improving workflow integration.
- Updated PlanLinker to utilize the new PlanCommitService for better phase tracking and commit management.
- Enhanced the PlanLinkerParser for improved parsing of plan files and extraction of relevant metadata.
- Refactored orchestration components to streamline agent and documentation linking based on updated registries.
- Added tests for the DocumentLinker to ensure proper mapping of documentation to agents and phases.
